fixed  the feedback that I was given
