+++ .env
@@ -1,0 +1,1 @@
+DB_CONNECT = 'postgressql://postgres:123@localhost:6000/papel'
