+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules
 bin
+draft.js
