+++ package.json
@@ -6,6 +6,7 @@
   "extends": "airbnb",
   "dependencies": {
     "bcrypt": "^3.0.8",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "pg": "^7.18.1"
   },
